+++ test/agent.js
@@ -11,16 +11,16 @@ var request = require('../lib/request')
 var Agent = require('../lib/agent')
 
 var opts = {
-  appName: 'some-app-name',
+  serviceName: 'some-service-name',
   secretToken: 'secret',
   captureExceptions: false,
   logLevel: 'error'
 }
 
 var optionFixtures = [
-  ['appName', 'APP_NAME'],
+  ['serviceName', 'SERVICE_NAME'],
   ['secretToken', 'SECRET_TOKEN'],
-  ['appVersion', 'APP_VERSION'],
+  ['serviceVersion', 'SERVICE_VERSION'],
   ['logLevel', 'LOG_LEVEL', 'info'],
   ['hostname', 'HOSTNAME', os.hostname()],
   ['stackTraceLimit', 'STACK_TRACE_LIMIT', Infinity],
@@ -89,7 +89,7 @@ falsyValues.forEach(function (val) {
   test('should be disabled by envrionment variable ELASTIC_APM_ACTIVE set to: ' + util.inspect(val), function (t) {
     setup()
     process.env.ELASTIC_APM_ACTIVE = val
-    agent.start({ appName: 'foo', secretToken: 'baz' })
+    agent.start({ serviceName: 'foo', secretToken: 'baz' })
     t.equal(agent.active, false)
     delete process.env.ELASTIC_APM_ACTIVE
     t.end()
@@ -100,7 +100,7 @@ truthyValues.forEach(function (val) {
   test('should be enabled by envrionment variable ELASTIC_APM_ACTIVE set to: ' + util.inspect(val), function (t) {
     setup()
     process.env.ELASTIC_APM_ACTIVE = val
-    agent.start({ appName: 'foo', secretToken: 'baz' })
+    agent.start({ serviceName: 'foo', secretToken: 'baz' })
     t.equal(agent.active, true)
     delete process.env.ELASTIC_APM_ACTIVE
     t.end()
@@ -109,7 +109,7 @@ truthyValues.forEach(function (val) {
 
 test('should overwrite ELASTIC_APM_ACTIVE by option property active', function (t) {
   setup()
-  var opts = { appName: 'foo', secretToken: 'baz', active: false }
+  var opts = { serviceName: 'foo', secretToken: 'baz', active: false }
   process.env.ELASTIC_APM_ACTIVE = '1'
   agent.start(opts)
   t.equal(agent.active, false)
@@ -119,7 +119,7 @@ test('should overwrite ELASTIC_APM_ACTIVE by option property active', function (
 
 test('should default active to true if required options have been specified', function (t) {
   setup()
-  agent.start({ appName: 'foo', secretToken: 'baz' })
+  agent.start({ serviceName: 'foo', secretToken: 'baz' })
   t.equal(agent.active, true)
   t.end()
 })
@@ -169,23 +169,23 @@ test('should separate strings and regexes into their own blacklist arrays', func
   t.end()
 })
 
-test('missing appName => inactive', function (t) {
+test('missing serviceName => inactive', function (t) {
   setup()
   agent.start()
   t.equal(agent.active, false)
   t.end()
 })
 
-test('invalid appName => inactive', function (t) {
+test('invalid serviceName => inactive', function (t) {
   setup()
-  agent.start({appName: 'foo&bar'})
+  agent.start({serviceName: 'foo&bar'})
   t.equal(agent.active, false)
   t.end()
 })
 
-test('valid appName => active', function (t) {
+test('valid serviceName => active', function (t) {
   setup()
-  agent.start({appName: 'fooBAR0123456789_- '})
+  agent.start({serviceName: 'fooBAR0123456789_- '})
   t.equal(agent.active, true)
   t.end()
 })
@@ -320,7 +320,7 @@ test('#captureError()', function (t) {
   t.test('should merge context', function (t) {
     setup()
     agent.start({
-      appName: 'foo',
+      serviceName: 'foo',
       secretToken: 'baz',
       filterHttpHeaders: false,
       logLevel: 'error'